requirement.txt
argparse
matplotlib
numpy
opencv-python
scipy
torchvision
tensorboardX
SimpleITK
pandas
xlsxwriter
skimage